requirements.txt
python-dotenv
requests
discord